optimize query

+++ src/main/java/com/greywanchuang/rackmonitor/domain/PSU.java
@@ -2,8 +2,8 @@ package com.greywanchuang.rackmonitor.domain;
 
 public class PSU {
     private String health;
-    private int voltage;
-    private int current;
+    private float voltage;
+    private float current;
 
     public String getHealth() {
         return health;
@@ -13,19 +13,19 @@ public class PSU {
         this.health = health;
     }
 
-    public int getVoltage() {
+    public float getVoltage() {
         return voltage;
     }
 
-    public void setVoltage(int voltage) {
+    public void setVoltage(float voltage) {
         this.voltage = voltage;
     }
 
-    public int getCurrent() {
+    public float getCurrent() {
         return current;
     }
 
-    public void setCurrent(int current) {
+    public void setCurrent(float current) {
         this.current = current;
     }
 }
